# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Phantom API",
    "description": "Phantom API"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "name": "localhost"
    }
  ],
  "paths": {
...
    "/api/v1/auth/register":{
      "post": {
        "tags": ["User"],
        "summary": " Create a new  user",
        "description": "Create user",
        "operationId": "createUser",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "type": "string",
            "description": "token"
          },
          {
            "in": "body",
            "name": "user",
            "description": "input name,email,birthdate and gender",
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Valid inputs are required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "example@phantom.com"
          },
          "password": {
            "type": "string",
            "example": "***********"
          },
          "role": {
            "type": "string"
          },
          "birthdate": {
            "type": "integer",
            "description": "User Role",
            "format": "int32"
          },
          "gender": {
            "type": "string"
          },
          "createdAt": {
            "type": "integer",
            "description": "User Status",
            "format": "int32"
          },
          "UpdatedAt": {
            "type": "integer",
            "description": "User Status",
            "format": "int32"
          }
        },
        "xml": {
          "name": "User"
        }
      },
...
    }
  }
}